+++ src/components/presentation-nav/nav-step.js
@@ -19,8 +19,8 @@ export default class NavStep extends React.Component {
 
     render() {
         let {step, onClick, activeStep, progress } = this.props;
-        let active = (activeStep == step.name);
-        let disabled = (progress == 0 && step.step > 1);
+        let active = (activeStep === step.name);
+        let disabled = (progress === 0 && step.step > 1);
         let future = (!active && progress < step.step);
         let completed = (!active &&  progress >= step.step);
