+++ src/components/presentation-nav/presentation-nav.less
@@ -30,7 +30,7 @@ ul.presentation-nav-steps li {
   margin: 0px;
   padding: 0 2%;
   border-top: 2px solid #D3DEEA;
-  width: 25%;
+  width: 20%;
   text-align: center;
   border-right: 1px solid #D3DEEA;
   height: 47px;
