+++ src/components/presentation-summary-form.js
@@ -14,11 +14,10 @@
 import React from 'react'
 import T from 'i18n-react/dist/i18n-react'
 import 'awesome-bootstrap-checkbox/awesome-bootstrap-checkbox.css'
-import { Input, TextEditor, UploadInput, Dropdown, RadioList, TextArea, Exclusive } from 'openstack-uicore-foundation/lib/components'
+import { Input, TextEditor, Dropdown, RadioList, TextArea } from 'openstack-uicore-foundation/lib/components'
 import {findElementPos} from 'openstack-uicore-foundation/lib/methods'
 import SubmitButtons from './presentation-submit-buttons'
 import {validate, scrollToError} from '../utils/methods'
-import Swal from "sweetalert2";
 
 
 class PresentationSummaryForm extends React.Component {
@@ -32,9 +31,6 @@ class PresentationSummaryForm extends React.Component {
 
         this.handleChange = this.handleChange.bind(this);
         this.handleSubmit = this.handleSubmit.bind(this);
-        this.handleUploadFile = this.handleUploadFile.bind(this);
-        this.handleRemoveFile = this.handleRemoveFile.bind(this);
-        this.handleFileError = this.handleFileError.bind(this);
     }
 
     componentWillReceiveProps(nextProps) {
@@ -50,84 +46,6 @@ class PresentationSummaryForm extends React.Component {
             let firstNode = document.getElementById(firstError);
             if (firstNode) window.scrollTo(0, findElementPos(firstNode));
         }
-    }
-
-    handleUploadFile(file, props) {
-
-        let { errors, entity} = this.state;
-        let {mediatype, mediaupload } = props;
-
-        if(mediaupload == null){
-            // new media upload
-            mediaupload = {
-                id: 0,
-                media_upload_type : mediatype,
-            };
-            // add
-            entity.media_uploads = [...entity.media_uploads, mediaupload];
-        }
-
-        if(mediaupload.hasOwnProperty('should_delete'))
-            delete mediaupload.should_delete;
-
-        mediaupload.file = file;
-        mediaupload.private_url = file.preview;
-        mediaupload.filename = file.name;
-
-        // update
-        entity.media_uploads = entity.media_uploads.map((item, index) => {
-            if (index !== mediaupload.index) {
-                // This isn't the item we care about - keep it as-is
-                return item
-            }
-
-            return {
-                ...item,
-                ...mediaupload
-            }
-        });
-
-        delete errors[mediatype.name];
-
-        this.setState({...this.state, entity, errors});
-    }
-
-    handleFileError(error, props){
-        if(error.length > 0){
-            let file = error[0];
-            let {mediatype } = props;
-            Swal.fire("Validation Error", `File size is greather than allowed ${mediatype.max_size/1024} MB`, "warning");
-        }
-    }
-
-    handleRemoveFile(ev, props) {
-
-        let entity = {...this.state.entity};
-        let { mediaupload } = props;
-
-        mediaupload.should_delete = true;
-        mediaupload.private_url = '';
-        mediaupload.file = null;
-
-        if(mediaupload.id  > 0 ){
-            entity.media_uploads = entity.media_uploads.map((item, index) => {
-                if (index !== mediaupload.index) {
-                    // This isn't the item we care about - keep it as-is
-                    return item
-                }
-
-                return {
-                    ...item,
-                    ...mediaupload
-                }
-            });
-        }
-        else{
-            // delete it
-            entity.media_uploads=  [...entity.media_uploads.slice(0, mediaupload.index), ...entity.media_uploads.slice(mediaupload.index + 1)];
-        }
-
-        this.setState({entity:entity});
     }
 
     handleChange(ev) {
@@ -179,28 +97,8 @@ class PresentationSummaryForm extends React.Component {
 
         validate(entity, rules, errors);
 
-        var cur_event_type = summit.event_types.find(ev => ev.id === entity.type_id);
-        let allowed_media_upload_types = [];
-        if(cur_event_type){
-            allowed_media_upload_types = cur_event_type.allowed_media_upload_types;
-        }
-
-        for(var mediaUploadType of allowed_media_upload_types){
-            if(mediaUploadType.is_mandatory){
-                // check if user provided file
-                var mediaUpload = this.getMediaUploadByType(entity, mediaUploadType);
-                if(! mediaUpload){
-                    errors[mediaUploadType.name] = 'This field is required.';
-                }
-
-                if(this.getMediaUploadFile(entity, mediaUploadType) == null && this.getMediaUploadFilePreview(entity, mediaUploadType) === ''){
-                    errors[mediaUploadType.name] = 'This field is required.';
-                }
-            }
-        }
-
         if (Object.keys(errors).length === 0) {
-            this.props.onSubmit(entity, 'tags');
+            this.props.onSubmit(entity, 'uploads');
         } else {
             this.setState({errors}, () => {
                 if (Object.keys(errors).length > 0) {
@@ -219,50 +117,11 @@ class PresentationSummaryForm extends React.Component {
         return '';
     }
 
-    getMediaUploadFile(entity, mediaType){
-        if(entity.media_uploads.length > 0 ){
-            let mediaUpload = entity.media_uploads.filter(mu => mu.hasOwnProperty('media_upload_type') && mu.media_upload_type.id === mediaType.id);
-            if(mediaUpload.length > 0){
-                return mediaUpload[0].file;
-            }
-        }
-        return null;
-    }
-
-    getMediaUploadFilePreview(entity, mediaType){
-        if(entity.media_uploads.length > 0 ){
-            let mediaUpload = entity.media_uploads.filter(mu => mu.hasOwnProperty('media_upload_type') && mu.media_upload_type.id === mediaType.id);
-            if(mediaUpload.length > 0){
-                if(mediaUpload[0].hasOwnProperty('should_delete') && mediaUpload[0].should_delete)
-                    return '';
-
-                return mediaUpload[0].hasOwnProperty('private_url') && mediaUpload[0].private_url !== '' ?  mediaUpload[0].private_url : mediaUpload[0].public_url ;
-            }
-        }
-        return '';
-    }
-
-    getMediaUploadFileName(entity, mediaType){
-        if(entity.media_uploads.length > 0 ){
-            let mediaUpload = entity.media_uploads.filter(mu => mu.hasOwnProperty('media_upload_type') && mu.media_upload_type.id === mediaType.id);
-            if(mediaUpload.length > 0){
-                return mediaUpload[0].hasOwnProperty('should_delete') && mediaUpload[0].should_delete ? '' : mediaUpload[0].filename;
-            }
-        }
-        return '';
-    }
-
-    getMediaUploadByType(entity, mediaType){
-        if(entity.media_uploads.length > 0 )
-            return entity.media_uploads.find(mu => mu.hasOwnProperty('media_upload_type') && mu.media_upload_type.id === mediaType.id);
-        return null;
-    }
-
     render() {
         let {entity} = this.state;
         let {selectionPlan, summit, presentation, step} = this.props;
 
-        if (!summit || !selectionPlan) return(<div></div>);
+        if (!summit || !selectionPlan) return(<div/>);
 
         let event_types_ddl = summit.event_types
             .filter(et => et.should_be_available_on_cfp)
@@ -305,12 +164,6 @@ class PresentationSummaryForm extends React.Component {
             {label: T.translate("general.yes"), value: 1},
             {label: T.translate("general.no"), value: 0}
         ];
-
-        var cur_event_type = summit.event_types.find(ev => ev.id === entity.type_id);
-        let allowed_media_upload_types = [];
-        if(cur_event_type){
-            allowed_media_upload_types = cur_event_type.allowed_media_upload_types;
-        }
 
         return (
             <form className="presentation-summary-form">
@@ -419,43 +272,6 @@ class PresentationSummaryForm extends React.Component {
                         <Input className="form-control" id="link_4" data-key="4" value={entity.links[4]} onChange={this.handleChange} error={this.hasErrors('link_4')} />
                     </div>
                 </div>
-
-
-                { allowed_media_upload_types.length > 0 &&
-
-                    <Exclusive name="presentation-attachment">
-
-                        <hr/>
-                        {
-                            allowed_media_upload_types.map(media_type => (
-                                <div key={media_type.id} className="row form-group">
-                                    <div className="col-md-12">
-                                        <label>{media_type.name} ({media_type.type.allowed_extensions.map((ext) => `.${ext.toLowerCase()}`).join(",")}) - Max. Size {media_type.max_size/1024} MB</label>
-                                        {
-                                            media_type.description !== '' &&
-                                            <h4>{media_type.description}</h4>
-                                        }
-                                        <UploadInput
-                                            fileName={this.getMediaUploadFileName(entity, media_type)}
-                                            mediatype= {media_type}
-                                            mediaupload={this.getMediaUploadByType(entity, media_type)}
-                                            handleUpload={this.handleUploadFile}
-                                            handleRemove={this.handleRemoveFile}
-                                            handleError={this.handleFileError}
-                                            className="dropzone col-md-6"
-                                            multiple={false}
-                                            maxSize={media_type.max_size * 1024}
-                                            error={this.hasErrors(media_type.name)}
-                                            value={this.getMediaUploadFilePreview(entity, media_type)}
-                                            file={this.getMediaUploadFile(entity, media_type)}
-                                            accept={media_type.type.allowed_extensions.map((ext) => `.${ext.toLowerCase()}`).join(",")}
-                                        />
-                                    </div>
-                                </div>
-                            ))
-                        }
-                    </Exclusive>
-                }
                 <hr/>
                 <SubmitButtons presentation={presentation} step={step} onSubmit={this.handleSubmit.bind(this)} />
             </form>
