+++ src/i18n/zh.json
@@ -114,6 +114,7 @@
   },
   "presentation_nav": {
     "summary": "总结",
+    "uploads": "Uploads",
     "tags": "标签",
     "speakers": "演讲者",
     "review": "审核和提交"
